+++ CHANGELOG.md
@@ -164,7 +164,7 @@
 
 ## [3.1.29] - 2018-05-14
 *   fix(wm): fix genCardConfig when not found any video card
-*   fix(watchdog): dde-polkit-agent running state check wrong
+*   fix(watchdog): gxde-polkit-agent running state check wrong
 *   adjust `auto_launch.json`
 *   fix: launch group
 *   feat: wait wm launch